Working on Render by boundary. Works well, but three issues. - if you match surfaces it doesn't change the renders on the fly, have to switch out of and back into "render by boundary" - if you are in this mode and draw new sub-surfaces, they don't render properly. I have not touched the SubSurface.rb, I probably need to - when you switch to this mode, it shows you surface by surface as things change. I noticed "render by class" and "render by value" seem to pause, and then present it all at once. Ideally this would do the same. Maybe related, is that once you switch modes with "value" or "class" you can then quickly switch back and forth like it is cached. This doesn't happen with "by boundary"
- I happened to notice that any render mode switch in large file (like H shape office) takes a few minutes. I thought it would be nice to add the progress bar here.

note: as is I think this is functional, and ocudl be left in without a toolbar, but just via the menu option (plugins/openstudiorendering/by boundary). If we clean it up, then maybe add a toolbar. If we don't want this in, we could leave the code in and just remove the command from the menu bar, or we could revert and remove the code all together.